Scene Save
Somehow the scene I made had not saved correctly.  Thankfully, with the help of the TA, I was able to recover the terrain I had made and save it in a new scene (Lovely Landscape).

All the landforms, textures, trees, and grass/flowers appear to have been saved as a part of the terrain.  From here, I plan to re-add the water feature and then continue adding more new elements to the scene.