Additional Elements
I imported the Old Building model we looked at in class, and a couple of other assets from the Unity Asset Store.

I originally had a few cabins, a greenhouse, a water well, and even a lighthouse in addition to the rocks that I imported.  However, these new buildings and structures resulted in compiler errors, so I had to unimport them for now until I find out why or I find new assets that I can import that won't give me this error.