Building Physics Testing
Testing Rigidbodies and Box Colliders on the two buildings I have placed to try to find a way so that the player won't phase through the walls.  Unfortunately, all I've discovered currently is how to get buildings to fall through the ground or fly away when the player contacts them.  I reset them to base function and will try to figure this out tomorrow.