Rock Re-Adjust
To make things simpler and not have to manually add each rock and bog down my computer, I made it so I can add rocks to the environment in the same way you add trees and grass.  Most of the rocks are under the water, but I also have them scattered about the landscape for the player to encounter.