Rock Re-Adjust (REAL)
I realized that this is probably when I performed the actions I described previously  (re-adding the rocks).  The previous commit might be when I tried importing different building assets to put in the environment.  However, instead of just causing compiler errors like the last time I attempted to add new buildings, these caused Unity to crash on me.  I will ask the TA or professor for help regarding this.